Fix regression in v0.11.0 (#250)
* Fix regression in v0.11.0

Change the default for writer opt 'IgnoreNoName' to be
true. This was set to false, and causing the example to
not run properly.

A new option 'WithErrorOnNoName' has been added to be
used in cases when this behaviour is expected. The existing
option 'WithIgnoreNoName' has now been deprecated.

A simple bash-based acceptance test has been added to
prevent this issue in the future. This starts a registry,
compiles the example into a binary, and makes sure it
runs properly. This test is now run in CI.

Signed-off-by: Josh Dolitsky <josh@dolit.ski>

* Allow hostname override for CI

Signed-off-by: Josh Dolitsky <josh@dolit.ski>

* use plain http for example..

Signed-off-by: Josh Dolitsky <josh@dolit.ski>

* Add logic to wait for registry to come up

Signed-off-by: Josh Dolitsky <josh@dolit.ski>