oras manifest index: include artifactType in index descriptors
For each manifest in the generated index manifest that is of a recognized
media type and has an artifactType value, we'll copy that value into the
associated descriptor so that software consuming the index can optionally
ignore or reject descriptors of the wrong artifactType without having to
retrieve the full manifest first.

The behavior does not change for manifests of any other media type or
those which do not have an artifactType property.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>